requirements.txt
pandas
scikit-learn
streamlit>=1.32.0
matplotlib
seaborn
numpy
joblib
altair<5.0.0
python-dotenv

langgraph
langchain-groq
langchain-core

sentence-transformers
faiss-cpu

pydantic>=2.0
fpdf2
